+++ src/headless/converse-chatboxes.js
@@ -427,10 +427,13 @@ converse.plugins.add('converse-chatboxes', {
                  */
                 if (_converse.send_chat_state_notifications && this.get('chat_state')) {
                     _converse.api.send(
-                        $msg({'to':this.get('jid'), 'type': 'chat'})
-                            .c(this.get('chat_state'), {'xmlns': Strophe.NS.CHATSTATES}).up()
-                            .c('no-store', {'xmlns': Strophe.NS.HINTS}).up()
-                            .c('no-permanent-store', {'xmlns': Strophe.NS.HINTS})
+                        $msg({
+                            'id': _converse.connection.getUniqueId(),
+                            'to': this.get('jid'),
+                            'type': 'chat'
+                        }).c(this.get('chat_state'), {'xmlns': Strophe.NS.CHATSTATES}).up()
+                          .c('no-store', {'xmlns': Strophe.NS.HINTS}).up()
+                          .c('no-permanent-store', {'xmlns': Strophe.NS.HINTS})
                     );
                 }
             },
@@ -665,6 +668,21 @@ converse.plugins.add('converse-chatboxes', {
                 const chatbox = this.getChatBox(from_jid);
                 if (!chatbox) {
                     return true;
+                }
+                const id = message.getAttribute('id');
+                if (id) {
+                    const msg = chatbox.messages.findWhere({'msgid': id});
+                    if (!msg) {
+                        // This error refers to a message not included in our store.
+                        // We assume that this was a CSI message (which we don't store).
+                        // See https://github.com/conversejs/converse.js/issues/1317
+                        return;
+                    }
+                } else {
+                    // An error message without id likely means that we
+                    // sent a message without id (which shouldn't happen).
+                    _converse.log('Received an error message without id attribute!', Strophe.LogLevel.ERROR);
+                    _converse.log(message, Strophe.LogLevel.ERROR);
                 }
                 chatbox.createMessage(message, message);
                 return true;